// src/app/models/pmo.ts
export class pmo{
  project_code: string='';
  //project_name:string='';
  project_description: string='';
  project_type: string='';
  project_category: string='';
  project_status: string='';
  project_manager: string='';
  delivery_head: string='';
  management:string='';
  from_date: string='';
  to_date: string='';
  client_name: string='';
  project_region: string='';
  project_country: string='';
  delivery_model: string='';
  budgeted_hours: number=0;
  allocated_hours: number=0;
  worked_hours: number=0;
  notes: string='';
}